Revert "Fixing push to wrong branch"
