+++ SSH TunnelBuilder/AppDelegate.swift
@@ -9,8 +9,6 @@ import Cocoa
 
 @main
 class AppDelegate: NSObject, NSApplicationDelegate {
-    
-    @IBOutlet var fileMenu: NSMenu!
 
     func applicationDidFinishLaunching(_ aNotification: Notification) {
         // Insert code here to initialize your application
