+++ SSH TunnelBuilder/Assets.xcassets/AccentColor.colorset/Contents.json
@@ -1,33 +1,6 @@
 {
   "colors" : [
     {
-      "color" : {
-        "color-space" : "srgb",
-        "components" : {
-          "alpha" : "1.000",
-          "blue" : "0.999",
-          "green" : "0.478",
-          "red" : "0.004"
-        }
-      },
-      "idiom" : "universal"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "color" : {
-        "color-space" : "srgb",
-        "components" : {
-          "alpha" : "1.000",
-          "blue" : "0.341",
-          "green" : "0.337",
-          "red" : "0.337"
-        }
-      },
       "idiom" : "universal"
     }
   ],
