+++ SSH TunnelBuilder/CONNECTION_DATA.json
@@ -1,0 +1,1 @@
+[{"connectionId":1,"sshHost":"192.168.9.140","sshHostPort":22,"localPort":8443,"remoteServer":"10.150.55.1","remotePort":443,"connectionName":"NHA RedZone","userName":"simon","password":"","publicKey":""}]
