+++ SSH TunnelBuilder/NIOSSHClient/InteractivePasswordPromptDelegate.swift
@@ -35,15 +35,26 @@ final class InteractivePasswordPromptDelegate: NIOSSHClientUserAuthenticationDel
     }
 
     func nextAuthenticationType(availableMethods: NIOSSHAvailableUserAuthenticationMethods, nextChallengePromise: EventLoopPromise<NIOSSHUserAuthenticationOffer?>) {
-        // guard availableMethods.contains(.password) || availableMethods.contains(.publicKey) else {
-        guard availableMethods.contains(.password) else {
+        guard availableMethods.contains(.password) || availableMethods.contains(.publicKey) else {
             print("Error: password auth not supported")
             nextChallengePromise.fail(SSHClientError.passwordAuthenticationNotSupported)
             return
         }
 
         self.queue.async {
-            if self.privateKey == "" {
+            if self.privateKey == nil {
+//                print("Username: ", terminator: "")
+//                self.username = readLine() ?? ""
+//            }
+//
+//            if self.password == nil {
+//                #if os(Windows)
+//                print("Password: ", terminator: "")
+//                self.password = readLine() ?? ""
+//                #else
+//                self.password = String(cString: getpass("Password: "))
+//                #endif
+//            }
 
             nextChallengePromise.succeed(NIOSSHUserAuthenticationOffer(username: self.username!, serviceName: "", offer: .password(.init(password: self.password!))))
