+++ SSH TunnelBuilder/NIOSSHClient/SSHClient.swift
@@ -62,7 +62,7 @@ class SSHClient
         
         let bootstrap = ClientBootstrap(group: group)
                         .channelInitializer { channel in
-                            channel.pipeline.addHandlers([NIOSSHHandler(role: .client(.init(userAuthDelegate: InteractivePasswordPromptDelegate(username: connection.username, password: connectionPassword, privateKey: connection.privateKey), serverAuthDelegate: AcceptAllHostKeysDelegate())), allocator: channel.allocator, inboundChildChannelInitializer: nil), ErrorHandler()])
+                            channel.pipeline.addHandlers([NIOSSHHandler(role: .client(.init(userAuthDelegate: InteractivePasswordPromptDelegate(username: connection.userName, password: connectionPassword, privateKey: connection.publicKey), serverAuthDelegate: AcceptAllHostKeysDelegate())), allocator: channel.allocator, inboundChildChannelInitializer: nil), ErrorHandler()])
                         }
                         .channelOption(ChannelOptions.socket(SocketOptionLevel(SOL_SOCKET), SO_REUSEADDR), value: 1)
                         .channelOption(ChannelOptions.socket(SocketOptionLevel(IPPROTO_TCP), TCP_NODELAY), value: 1)
